requirements.txt
torch
torchvision
torchaudio
transformers
peft
accelerate
bitsandbytes
datasets
sentence-transformers
FlagEmbedding
faiss-cpu
pandas
numpy
tqdm
openai
anthropic
python-dotenv
ipykernel
